# lab3example/src/main/java/info/stepanoff/trsis/samples/rest/model/StudentDataWithoutID.java
package info.stepanoff.trsis.samples.rest.model;

import lombok.Getter;

@Getter
public class StudentDataWithoutID {
    private String fullName;
    private Integer taskOption;
    private Integer number;
    private Integer rating;
}